refactor: split provider.rs into per-operation modules
Split the 3456-line carina-provider-awscc/src/provider.rs into focused
sub-modules:

- provider/mod.rs: AwsccProvider struct, constants, schema config lookup
- provider/cloudcontrol.rs: Cloud Control API methods and retry/error handling
- provider/conversion.rs: DSL <-> AWS JSON value conversion
- provider/normalizer.rs: Enum identifier resolution and state hydration
- provider/operations.rs: High-level resource CRUD operations
- provider/update.rs: Update patch building and property parsing

All public APIs remain accessible through the same paths. Tests are
co-located with their respective modules.

Part of #777
